# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Calculation": {
        "additionalProperties": false,
        "description": "Persisted as ``calculations/<id>/calculation.json`` inside the project.",
        "properties": {
          "backend_id": {
            "title": "Backend Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "generated": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GeneratedInputs"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "job": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "parent_calculation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "for reruns/restarts",
            "title": "Parent Calculation Id"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Provenance"
              },
              {
                "type": "null"
              }
            ]
          },
          "resources": {
            "$ref": "#/components/schemas/Resources"
          },
          "result_structure_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Structure Id"
          },
          "results": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResultBundle"
              },
              {
                "type": "null"
              }
            ]
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "default": "draft",
            "enum": [
              "draft",
              "ready",
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "structure_id": {
            "description": "input structure (a copy is stored with the calculation)",
            "title": "Structure Id",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "description": "full merged parameter values",
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "name",
          "backend_id",
          "structure_id"
        ],
        "title": "Calculation",
        "type": "object"
      },
...
      "CreateCalculationRequest": {
        "additionalProperties": false,
        "properties": {
          "backend_id": {
            "title": "Backend Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/Resources"
          },
          "structure_id": {
            "title": "Structure Id",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "name",
          "backend_id",
          "structure_id"
        ],
        "title": "CreateCalculationRequest",
        "type": "object"
      },
...
      "GeneratedInputs": {
        "additionalProperties": false,
        "properties": {
          "files": {
            "items": {
              "$ref": "#/components/schemas/GeneratedFile"
            },
            "title": "Files",
            "type": "array"
          },
          "root_name": {
            "description": "base name of the calculation, e.g. 'case'",
            "title": "Root Name",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "files",
          "root_name"
        ],
        "title": "GeneratedInputs",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobRecord": {
        "additionalProperties": false,
        "properties": {
          "calculation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Calculation Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "exit_code": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exit Code"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "spec": {
            "$ref": "#/components/schemas/RunSpec"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "default": "queued",
            "enum": [
              "queued",
              "running",
              "completed",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "spec"
        ],
        "title": "JobRecord",
        "type": "object"
      },
      "LogResponse": {
        "additionalProperties": false,
        "properties": {
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          },
          "stream": {
            "title": "Stream",
            "type": "string"
          }
        },
        "required": [
          "stream",
          "lines"
        ],
        "title": "LogResponse",
        "type": "object"
      },
...
      "ParameterSchema": {
        "additionalProperties": false,
        "properties": {
          "backend": {
            "title": "Backend",
            "type": "string"
          },
          "id": {
            "description": "e.g. 'cppaw'",
            "title": "Id",
            "type": "string"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "title": "Sections",
            "type": "array"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "backend",
          "title",
          "sections"
        ],
        "title": "ParameterSchema",
        "type": "object"
      },
...
      "Preset": {
        "additionalProperties": false,
        "properties": {
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "schema_id": {
            "title": "Schema Id",
            "type": "string"
          },
          "values": {
            "additionalProperties": true,
            "description": "partial: only keys that differ",
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "id",
          "name",
          "schema_id"
        ],
        "title": "Preset",
        "type": "object"
      },
...
      "ResultBundle": {
        "additionalProperties": false,
        "description": "Parsed results in Atomscope model types. Backend-specific data stays in ``extra``.",
        "properties": {
          "converged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Converged"
          },
          "extra": {
            "additionalProperties": true,
            "title": "Extra",
            "type": "object"
          },
          "final_structure": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Structure"
              },
              {
                "type": "null"
              }
            ]
          },
          "grids": {
            "items": {
              "$ref": "#/components/schemas/VolumetricGrid"
            },
            "title": "Grids",
            "type": "array"
          },
          "properties": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Quantity"
            },
            "title": "Properties",
            "type": "object"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/ScalarSeries"
            },
            "title": "Series",
            "type": "array"
          },
          "trajectory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Trajectory"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "ResultBundle",
        "type": "object"
      },
...
      "Section": {
        "additionalProperties": false,
        "properties": {
          "advanced": {
            "default": false,
            "title": "Advanced",
            "type": "boolean"
          },
          "help": {
            "default": "",
            "title": "Help",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "parameters": {
            "items": {
              "$ref": "#/components/schemas/ParameterSpec"
            },
            "title": "Parameters",
            "type": "array"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "Section",
        "type": "object"
      },
...
      "UpdateValuesRequest": {
        "additionalProperties": false,
        "properties": {
          "values": {
            "additionalProperties": true,
            "title": "Values",
            "type": "object"
          }
        },
        "required": [
          "values"
        ],
        "title": "UpdateValuesRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "ValidationReport": {
        "additionalProperties": false,
        "properties": {
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ValidationIssue"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "title": "ValidationReport",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Atomscope API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/backends/{backend_id}/presets": {
      "get": {
        "operationId": "backend_presets_api_backends__backend_id__presets_get",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Preset"
                  },
                  "title": "Response Backend Presets Api Backends  Backend Id  Presets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backend Presets",
        "tags": [
          "backends"
        ]
      }
    },
    "/api/backends/{backend_id}/schema": {
      "get": {
        "operationId": "backend_schema_api_backends__backend_id__schema_get",
        "parameters": [
          {
            "in": "path",
            "name": "backend_id",
            "required": true,
            "schema": {
              "title": "Backend Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParameterSchema"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backend Schema",
        "tags": [
          "backends"
        ]
      }
    },
    "/api/calculations": {
      "get": {
        "operationId": "list_calculations_api_calculations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Calculation"
                  },
                  "title": "Response List Calculations Api Calculations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Calculations",
        "tags": [
          "calculations"
        ]
      },
      "post": {
        "operationId": "create_calculation_api_calculations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCalculationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Calculation",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}": {
      "get": {
        "operationId": "get_calculation_api_calculations__calc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Calculation",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/cancel": {
      "post": {
        "operationId": "cancel_calculation_api_calculations__calc_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Calculation",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/generate": {
      "post": {
        "operationId": "generate_inputs_api_calculations__calc_id__generate_post",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeneratedInputs"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Inputs",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/log": {
      "get": {
        "description": "Return the last ``tail`` lines of stdout/stderr or a watched file from the work directory.",
        "operationId": "read_log_api_calculations__calc_id__log_get",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "stream",
            "required": false,
            "schema": {
              "default": "stdout",
              "title": "Stream",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tail",
            "required": false,
            "schema": {
              "default": 500,
              "title": "Tail",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Log",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/results": {
      "get": {
        "operationId": "results_api_calculations__calc_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultBundle"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Results",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/run": {
      "post": {
        "description": "Async so that JobManager.submit runs inside the server's event loop.",
        "operationId": "run_calculation_api_calculations__calc_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Calculation",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/validate": {
      "get": {
        "operationId": "validate_calculation_api_calculations__calc_id__validate_get",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Calculation",
        "tags": [
          "calculations"
        ]
      }
    },
    "/api/calculations/{calc_id}/values": {
      "put": {
        "operationId": "update_values_api_calculations__calc_id__values_put",
        "parameters": [
          {
            "in": "path",
            "name": "calc_id",
            "required": true,
            "schema": {
              "title": "Calc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Calculation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Values",
        "tags": [
          "calculations"
        ]
      }
    },
...
  }
}